backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.4.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-rest-passwordreset==1.5.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
idna==3.10
pycryptodome==3.21.0
PyJWT==2.9.0
PyMySQL==1.1.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
cryptography==41.0.3
gunicorn==20.1.0

